refactor project and take payload as module scoped variable
